Threading: The syncs are transitive, so use this info to allow for more cleaning
In the event that a thread t1 will sync with another thread t2, then we can
compare the other thread syncs to allow for more deletions. Consider this:
t2 has synced with a third thread t3 at a later state than t1. In that event
we know that the sync t2 <-> t3 happend before t1 <-> t3. Since now t1 <-> t2
happend, we can definetly say that the sync with t3 also applies to t1