fix: store threadId for MemoryObjects
Corrects a bug where for certain local objects that were allocated
in one thread, but modified in an other thread, the wrong delta of
the fingerprint is modified. This can lead to cancellation and
was detected by an assertion failure in VerifiedMemoryFingerprint.

An example is a pthread_t that was allocated in main() and is
modified at the start of a new thread. Then, the index of the
stackframe is 0, but the threadId is different from the current
one, which is where the delta was wrongly modified.