guard stack access for exited threads
Executor::processTimers():
Quick fix, which might not be adequate for some usage of dumpStates.

StatsTracker::updateStateStatistics():
Do not update stats for states currently executing an exited thread.
This method is only relevant for writeIStats.

Searcher::getWeight():
Give maximal weight whenever we are currently executing a state that has exited
to eliminate the risk of this state never seeing daylight again. The assumption
behind this: Such an ExecutionState will either near its end or execute
something different on selection (at which point we can apply the search
heuristics as per usual, again).