Fix key type confusion
Since we do not want to expose the `kpr_list` type in the `pthread.h` header
we simply relied on the trick to let the `pthread_key_t` type be a void pointer.
With that we could hide the underlying type.
But there is one problem: the compiler will often not issue a warning or error
if the wrong type is supplied as an argument.
This is very important since the pthread_key_* api wants for the "create" a
pointer to the key and for the other calls simply the `pthread_key_t` type.